+++ schools/forms.py
@@ -1,0 +1,5 @@
+from django                     import forms
+from django.forms.widgets       import SelectDateWidget
+from django.utils.translation   import ugettext_lazy as _
+
+from common.models import *
